+++ CHANGELOG.md
@@ -8,6 +8,13 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * The MINOR component is used whenever a version has backwards-compatible profile changes. This also indicates that the user can not switch back to a previous MINOR version without using a backup.
 * The PATCH component is used for versions that do not contain profile format changes. Users can freely switch between PATCH versions without risk of data loss.
 
+## [Unreleased]
+
+### Added
+
+- Add support for Wrath of the Lich King Classic 3.4.0
+- Add talent selected load condtion for WOTLK
+
 ## [1.7.4] - 2022-08-25
 
 ### Fixed